+++ frontend/app/(components)/Profile/Update/Update.module.css
@@ -1,0 +1,75 @@
+.profileSection {
+    flex: 0 0 300px;
+    background-color: #2c2c2c;
+    /*background-color: var(--background);*/
+
+    border-radius: 10px;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    overflow-y: auto;
+    max-height: calc(100vh - 40px);
+}
+
+.profileImage {
+    width: 150px;
+    height: 150px;
+    border-radius: 50%;
+    background-color: #3a3a3a;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    overflow: hidden;
+    margin-bottom: 20px;
+}
+
+.profileImageContent {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.nickname {
+    font-size: 1.2em;
+    font-weight: bold;
+    margin-bottom: 20px;
+}
+
+.button {
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 10px;
+    background-color: #3a3a3a;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
+}
+
+.button:hover {
+    background-color: #4a4a4a;
+}
+
+.editForm {
+    width: 100%;
+    max-height: 0;
+    overflow: hidden;
+    transition: max-height 0.3s ease-out;
+}
+
+.editForm.visible {
+    max-height: 1000px;
+    transition: max-height 0.5s ease-in;
+}
+
+.input {
+    width: 100%;
+    padding: 10px;
+    margin-bottom: 10px;
+    /*프로필수정쪽 텍스트칸 컬러*/
+    background-color: #3a3a3a;
+    color: white;
+    border: none;
+    border-radius: 5px;
+}
