+++ frontend/app/(components)/Profile/profile.module.css
@@ -18,20 +18,6 @@
   margin-left: 60px;
 }
 
-.profileSection {
-  flex: 0 0 300px;
-  background-color: #2c2c2c;
-  /*background-color: var(--background);*/
-
-  border-radius: 10px;
-  padding: 20px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  overflow-y: auto;
-  max-height: calc(100vh - 40px);
-}
-
 .contentSection {
   flex: 1;
   display: flex;
@@ -40,129 +26,3 @@
   overflow-y: auto;
   max-height: calc(100vh - 40px);
 }
-
-.section {
-  /*내가남긴리뷰 좋아요 누른 영화 부분 색상*/
-  background-color: #2c2c2c;
-  /*background-color: var(--background);*/
-
-  border-radius: 10px;
-  padding: 20px;
-  flex: 1;
-}
-
-.profileImage {
-  width: 150px;
-  height: 150px;
-  border-radius: 50%;
-  background-color: #3a3a3a;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-  margin-bottom: 20px;
-}
-
-.profileImageContent {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.nickname {
-  font-size: 1.2em;
-  font-weight: bold;
-  margin-bottom: 20px;
-}
-
-.button {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 10px;
-  background-color: #3a3a3a;
-  color: white;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
-
-.button:hover {
-  background-color: #4a4a4a;
-}
-
-.input {
-  width: 100%;
-  padding: 10px;
-  margin-bottom: 10px;
-  /*프로필수정쪽 텍스트칸 컬러*/
-  background-color: #3a3a3a;
-  color: white;
-  border: none;
-  border-radius: 5px;
-}
-
-.editForm {
-  width: 100%;
-  max-height: 0;
-  overflow: hidden;
-  transition: max-height 0.3s ease-out;
-}
-
-.editForm.visible {
-  max-height: 1000px;
-  transition: max-height 0.5s ease-in;
-}
-
-.sectionTitle {
-  text-align: center;
-  margin-bottom: 10px;
-  font-weight: bold;
-}
-
-.error {
-  color: red;
-  font-size: 0.9em;
-  margin-bottom: 10px;
-}
-
-.success {
-  color: green;
-  font-size: 0.9em;
-  margin-bottom: 10px;
-}
-
-.navButton {
-  padding: 10px;
-  background-color: transparent;
-  color: #007bff;
-  border: none;
-  border-radius: 50%;
-  cursor: pointer;
-  font-size: 2rem;
-}
-.movieItems {
-  list-style-type: none;
-  padding: 0;
-  margin: 0; /* 마진을 0으로 설정하여 불필요한 간격 제거 */
-  display: flex;
-  transition: transform 0.5s ease; /* 슬라이드 애니메이션 */
-}
-
-.movieItem {
-  box-sizing: border-box;
-  flex: 0 0 200px; /* 각 포스터의 너비를 200px로 고정 */
-  margin: 0 10px; /* 포스터 좌우 간격 추가 */
-}
-
-.movieImg {
-  width: 200px; /* 고정된 너비 */
-  height: 300px; /* 고정된 높이 */
-  object-fit: cover; /* 이미지를 셀에 맞게 자름 */
-  transition: transform 0.2s ease;
-}
-
-.sliderWrapper {
-  overflow: hidden;
-  width: 1100px; /* 포스터 5개의 너비에 맞게 고정 */
-}
-
